keep a refactor for later to add coverage to all AST related
